메인으로 카드 함수 옮겨 모달 전역관리 세팅
