+++ src/pages/Main/Main.jsx
@@ -8,6 +8,12 @@ import { POSTBOXES_API } from 'config';
 import { useInView } from 'react-intersection-observer';
 import { Card } from 'components/Card';
 
+import { modalState, actions } from 'atom';
+import { chkPwd } from 'Validation/Validation';
+import { useRecoilState } from 'recoil';
+import { Modal } from 'components/Modal';
+import { SendingForm, PwForm } from 'components/Card/Form';
+
 const Main = () => {
   const [sortList, setSortList] = useState('');
 
@@ -20,6 +26,19 @@ const Main = () => {
 
   const [ref, inView] = useInView();
 
+  //모달 용 state
+  const [isModal, setIsModal] = useRecoilState(modalState);
+  const [selectedFiles, setSelectedFiles] = useState(null);
+  const [formValues, setFormValues] = useState({
+    nameInput: '',
+    textInput: '',
+    pwInput: '',
+    boxId: '',
+  });
+
+  console.log(`letterBoxList`, letterBoxList);
+  //
+
   const getItems = useCallback(async () => {
     setLoading(true);
     await axios
@@ -29,7 +48,6 @@ const Main = () => {
         setLoading(false);
       });
   }, [page, sortList, trigger]);
-  // console.log(`sortList`, sortList);
   // `getItems` 가 바뀔 때 마다 함수 실행
   useEffect(() => {
     getItems();
@@ -41,6 +59,7 @@ const Main = () => {
       setPage(prevState => prevState + 1);
     }
   }, [inView, loading]);
+  console.log(`loading`, loading);
 
   const sortClicked = e => {
     const filterType = e.target.innerText;
@@ -60,6 +79,101 @@ const Main = () => {
     setFilterType(filterTypeToEng[filterType]);
   };
 
+  //Card 모달 함수
+  const openModal = () => {
+    setFormValues({ boxId: letterBoxList.id });
+    if (letterBoxList.is_public === true) {
+      setIsModal(actions.OPEN_SEND);
+      // setModalOpen(true);
+      document.body.style.overflow = 'hidden';
+    }
+    if (letterBoxList.is_public === false) {
+      setIsModal(actions.OPEN_PW);
+      // setOpenPw(true);
+      document.body.style.overflow = 'hidden';
+    }
+  };
+  //메일 전송 완료 시 닫는 함수
+  const closeModalState = () => {
+    setIsModal(actions.CLOSE_SEND);
+    setIsModal(actions.CLOSE_PW);
+    document.body.style.overflow = 'unset';
+  };
+
+  //이번트 겟
+  const handleForm = e => {
+    const { name, value } = e.target;
+    setFormValues({ ...formValues, [name]: value });
+  };
+
+  //파일 업로드
+  const fileChangedHandler = e => {
+    setSelectedFiles(e.target.files[0]);
+  };
+  // 이메일 보내기 // to collection
+  const sendMail = () => {
+    let token = localStorage.getItem('TOKEN') || '';
+    const formData = new FormData();
+    formData.append('image', selectedFiles);
+    formData.append('nickname', formValues.nameInput);
+    formData.append('caption', formValues.textInput);
+    const config = {
+      headers: {
+        Authorization: token,
+        'content-type': 'multipart/form-data',
+      },
+    };
+    if (
+      formValues.nameInput &&
+      formValues.textInput &&
+      selectedFiles !== null
+    ) {
+      axios.post(`${POSTBOXES_API}/${formValues.boxId}/send`, formData, config);
+      //  .then(res => res.json())
+      // .then(res => {
+      //   {
+      //     localStorage.setItem('message', res.data.message);
+      //   }
+      // });
+      alert('전송완료');
+      setSelectedFiles(null);
+      localStorage.removeItem('TOKEN');
+      setFormValues({ nameInput: '', textInput: '' });
+      closeModalState();
+    } else alert('정해진 양식을 채워주세요.');
+  };
+  //비밀번호 체크
+  const checkPw = () => {
+    if (chkPwd(formValues.pwInput)) {
+      fetch(`${POSTBOXES_API}/access`, {
+        method: 'POST',
+        headers: {
+          'content-type': 'application/json',
+        },
+        body: JSON.stringify({
+          id: formValues.boxId,
+          password: formValues.pwInput,
+        }),
+      })
+        .then(res => res.json())
+        .then(res => {
+          if (res.token) {
+            localStorage.setItem('TOKEN', res.token);
+            // this.props.history.push('/');
+            // setModalOpen(true);
+            // setOpenPw(false);
+            setIsModal(actions.OPEN_SEND);
+            setIsModal(actions.CLOSE_PW);
+
+            setFormValues({ ...formValues, pwInput: '' });
+          } else {
+            alert('입력하신 비밀번호를 다시 확인해주세요.');
+          }
+        });
+    } else {
+      alert('숫자와 영문자 조합으로 8~15자리를 사용해야 합니다.');
+    }
+  };
   // '' public private
   return (
     <Container>
@@ -76,10 +190,42 @@ const Main = () => {
           <React.Fragment key={idx}>
             {letterBoxList.length - 1 === idx ? (
               <div ref={ref}>
-                <Card key={idx} letterBox={letterBox} />
+                <Card key={idx} letterBox={letterBox} openModal={openModal}>
+                  {/* 모달창 */}
+                  {isModal.type === 'checkPw' && (
+                    <Modal header="비밀번호">
+                      <PwForm handleForm={handleForm} checkPw={checkPw} />
+                    </Modal>
+                  )}
+                  {isModal.type === 'sendMail' && (
+                    <Modal header="이메일 보내기">
+                      <SendingForm
+                        handleForm={handleForm}
+                        fileChangedHandler={fileChangedHandler}
+                        sendMail={sendMail}
+                      />
+                    </Modal>
+                  )}
+                </Card>
               </div>
             ) : (
-              <Card key={idx} letterBox={letterBox} />
+              <Card key={idx} letterBox={letterBox} openModal={openModal}>
+                {/* 모달창 */}
+                {isModal.type === 'checkPw' && (
+                  <Modal header="비밀번호">
+                    <PwForm handleForm={handleForm} checkPw={checkPw} />
+                  </Modal>
+                )}
+                {isModal.type === 'sendMail' && (
+                  <Modal header="이메일 보내기">
+                    <SendingForm
+                      handleForm={handleForm}
+                      fileChangedHandler={fileChangedHandler}
+                      sendMail={sendMail}
+                    />
+                  </Modal>
+                )}
+              </Card>
             )}
           </React.Fragment>
         ))}
